[rollout, vllm] test: merge the two vision placeholder test files into one
Per review. Same 7 cases, now grouped into which ids to ban and where they
get masked, so the two halves of one behaviour read as one file.

Dropping the mask line still turns test_banned_tokens_and_oov_tail_are_masked
red, so the merge did not cost the coverage.
